balancing and info screen
made everything a bit faster, incl player. tried to rebalance the towers to be more suited for their scrap cost, as well as made enemies in general have some more health and adjusted. the info screen with tower and enemy info exists but not hooked up to anything just yet.